Wrap autossh in an infinite loop (#20)
Autossh will exit if the first invocation of SSH fails which can happen if the host restarts (because of power failure?) and network isn't yet available (or your are unlucky and your tunnel target host is unavailable during this critical period). This causes the whole add-on to terminate which isn't great if you are relying on it to maintain access to a remote system. This commit makes autossh run in an infinite loop to ensure it never exits by itself.

Co-authored-by: Andre Borie <hi@rjevski.io>